koishi/finance/tempo/code/requirements.txt
yfinance
psycopg2-binary
pytz
pandas
